requirements.txt
openai
feedparser
python-dotenv
requests